Complete rework of the repo, split it up to support multiple syntax styles and different frameworks
